pyproject.toml
[project]
name = "steam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cookiecutter",
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "websockets",
    "invoke",
    "libsass",
]
